requirements.txt
aiosqlite==0.20.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.28.0
anyio==4.12.1
APScheduler==3.10.4
beautifulsoup4==4.14.3
boto3==1.42.73
botocore==1.42.73
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
curl_cffi==0.13.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.24
feedparser==6.0.11
filelock==3.25.2
frozendict==2.4.7
fsspec==2026.2.0
google-ai-generativelanguage==0.6.6
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.7.2
googleapis-common-protos==1.73.0
greenlet==3.3.2
groq==0.9.0
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.4.2
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
metaflow==2.11.0
mpmath==1.3.0
multitasking==0.0.12
networkx==3.6.1
nltk==3.9.3
nsepython==2.97
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
orjson==3.11.7
packaging==26.0
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
peewee==4.0.2
pillow==12.1.1
platformdirs==4.9.4
playwright==1.58.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pyparsing==3.3.2
pypdfium2==5.6.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.31.0
rich==14.3.3
rich-toolkit==0.19.7
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==3.0.1
setuptools==82.0.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.37.2
sympy==1.14.0
TextBlob==0.18.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==4.57.6
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
webencodings==0.5.1
websockets==16.0
yfinance==1.2.0
